capabilities_server: move some struct to capabilities_server
