+++ general_gojay.go
@@ -207,31 +207,6 @@ func (v *FileOperationFilters) UnmarshalJSONArray(dec *gojay.Decoder) error {
 }
 
 // MarshalJSONObject implements gojay.MarshalerJSONObject.
-func (v *FileOperationRegistrationOptions) MarshalJSONObject(enc *gojay.Encoder) {
-	enc.ArrayKey(keyFilters, FileOperationFilters(v.Filters))
-}
-
-// IsNil implements gojay.MarshalerJSONObject.
-func (v *FileOperationRegistrationOptions) IsNil() bool { return v == nil }
-
-// UnmarshalJSONObject implements gojay.UnmarshalerJSONObject.
-func (v *FileOperationRegistrationOptions) UnmarshalJSONObject(dec *gojay.Decoder, k string) error {
-	if k == keyFilters {
-		return dec.Array((*FileOperationFilters)(&v.Filters))
-	}
-	return nil
-}
-
-// NKeys implements gojay.UnmarshalerJSONObject.
-func (v *FileOperationRegistrationOptions) NKeys() int { return 1 }
-
-// compile time check whether the FileOperationRegistrationOptions implements a gojay.MarshalerJSONObject and gojay.UnmarshalerJSONObject interfaces.
-var (
-	_ gojay.MarshalerJSONObject   = (*FileOperationRegistrationOptions)(nil)
-	_ gojay.UnmarshalerJSONObject = (*FileOperationRegistrationOptions)(nil)
-)
-
-// MarshalJSONObject implements gojay.MarshalerJSONObject.
 func (v *FileOperationPatternOptions) MarshalJSONObject(enc *gojay.Encoder) {
 	enc.BoolKeyOmitEmpty(keyIgnoreCase, v.IgnoreCase)
 }
@@ -646,31 +621,6 @@ func (v *MarkupKinds) UnmarshalJSONArray(dec *gojay.Decoder) error {
 }
 
 // MarshalJSONObject implements gojay.MarshalerJSONObject.
-func (v *DocumentHighlightOptions) MarshalJSONObject(enc *gojay.Encoder) {
-	enc.BoolKeyOmitEmpty(keyWorkDoneProgress, v.WorkDoneProgress)
-}
-
-// IsNil returns wether the structure is nil value or not.
-func (v *DocumentHighlightOptions) IsNil() bool { return v == nil }
-
-// UnmarshalJSONObject implements gojay's UnmarshalerJSONObject.
-func (v *DocumentHighlightOptions) UnmarshalJSONObject(dec *gojay.Decoder, k string) error {
-	if k == keyWorkDoneProgress {
-		return dec.Bool(&v.WorkDoneProgress)
-	}
-	return nil
-}
-
-// NKeys returns the number of keys to unmarshal.
-func (v *DocumentHighlightOptions) NKeys() int { return 1 }
-
-// compile time check whether the DocumentHighlightOptions implements a gojay.MarshalerJSONObject and gojay.UnmarshalerJSONObject interfaces.
-var (
-	_ gojay.MarshalerJSONObject   = (*DocumentHighlightOptions)(nil)
-	_ gojay.UnmarshalerJSONObject = (*DocumentHighlightOptions)(nil)
-)
-
-// MarshalJSONObject implements gojay.MarshalerJSONObject.
 func (v *DocumentHighlightParams) MarshalJSONObject(enc *gojay.Encoder) {
 	enc.ObjectKeyOmitEmpty(keyTextDocument, &v.TextDocument)
 	enc.ObjectKeyOmitEmpty(keyPosition, &v.Position)
@@ -703,116 +653,6 @@ func (v *DocumentHighlightParams) NKeys() int { return 4 }
 var (
 	_ gojay.MarshalerJSONObject   = (*DocumentHighlightParams)(nil)
 	_ gojay.UnmarshalerJSONObject = (*DocumentHighlightParams)(nil)
-)
-
-// MarshalJSONObject implements gojay.MarshalerJSONObject.
-func (v *WorkspaceSymbolOptions) MarshalJSONObject(enc *gojay.Encoder) {
-	enc.BoolKeyOmitEmpty(keyWorkDoneProgress, v.WorkDoneProgress)
-}
-
-// IsNil implements gojay.MarshalerJSONObject.
-func (v *WorkspaceSymbolOptions) IsNil() bool { return v == nil }
-
-// UnmarshalJSONObject implements gojay.UnmarshalerJSONObject.
-func (v *WorkspaceSymbolOptions) UnmarshalJSONObject(dec *gojay.Decoder, k string) error {
-	if k == keyWorkDoneProgress {
-		return dec.Bool(&v.WorkDoneProgress)
-	}
-	return nil
-}
-
-// NKeys implements gojay.UnmarshalerJSONObject.
-func (v *WorkspaceSymbolOptions) NKeys() int { return 1 }
-
-// compile time check whether the WorkspaceSymbolOptions implements a gojay.MarshalerJSONObject and gojay.UnmarshalerJSONObject interfaces.
-var (
-	_ gojay.MarshalerJSONObject   = (*WorkspaceSymbolOptions)(nil)
-	_ gojay.UnmarshalerJSONObject = (*WorkspaceSymbolOptions)(nil)
-)
-
-// MarshalJSONObject implements gojay.MarshalerJSONObject.
-func (v *DocumentFormattingOptions) MarshalJSONObject(enc *gojay.Encoder) {
-	enc.BoolKeyOmitEmpty(keyWorkDoneProgress, v.WorkDoneProgress)
-}
-
-// IsNil returns wether the structure is nil value or not.
-func (v *DocumentFormattingOptions) IsNil() bool { return v == nil }
-
-// UnmarshalJSONObject implements gojay's UnmarshalerJSONObject.
-func (v *DocumentFormattingOptions) UnmarshalJSONObject(dec *gojay.Decoder, k string) error {
-	if k == keyWorkDoneProgress {
-		return dec.Bool(&v.WorkDoneProgress)
-	}
-	return nil
-}
-
-// NKeys returns the number of keys to unmarshal.
-func (v *DocumentFormattingOptions) NKeys() int { return 1 }
-
-// compile time check whether the DocumentFormattingOptions implements a gojay.MarshalerJSONObject and gojay.UnmarshalerJSONObject interfaces.
-var (
-	_ gojay.MarshalerJSONObject   = (*DocumentFormattingOptions)(nil)
-	_ gojay.UnmarshalerJSONObject = (*DocumentFormattingOptions)(nil)
-)
-
-// MarshalJSONObject implements gojay.MarshalerJSONObject.
-func (v *DeclarationOptions) MarshalJSONObject(enc *gojay.Encoder) {
-	enc.BoolKeyOmitEmpty(keyWorkDoneProgress, v.WorkDoneProgress)
-}
-
-// IsNil implements gojay.MarshalerJSONObject.
-func (v *DeclarationOptions) IsNil() bool { return v == nil }
-
-// UnmarshalJSONObject implements gojay.UnmarshalerJSONObject.
-func (v *DeclarationOptions) UnmarshalJSONObject(dec *gojay.Decoder, k string) error {
-	if k == keyWorkDoneProgress {
-		return dec.Bool(&v.WorkDoneProgress)
-	}
-	return nil
-}
-
-// NKeys implements gojay.UnmarshalerJSONObject.
-func (v *DeclarationOptions) NKeys() int { return 1 }
-
-// compile time check whether the DeclarationOptions implements a gojay.MarshalerJSONObject and gojay.UnmarshalerJSONObject interfaces.
-var (
-	_ gojay.MarshalerJSONObject   = (*DeclarationOptions)(nil)
-	_ gojay.UnmarshalerJSONObject = (*DeclarationOptions)(nil)
-)
-
-// MarshalJSONObject implements gojay.MarshalerJSONObject.
-func (v *DeclarationRegistrationOptions) MarshalJSONObject(enc *gojay.Encoder) {
-	enc.BoolKeyOmitEmpty(keyWorkDoneProgress, v.WorkDoneProgress)
-	enc.AddArrayKey(keyDocumentSelector, &v.DocumentSelector)
-	enc.StringKeyOmitEmpty(keyID, v.ID)
-}
-
-// IsNil implements gojay.MarshalerJSONObject.
-func (v *DeclarationRegistrationOptions) IsNil() bool { return v == nil }
-
-// UnmarshalJSONObject implements gojay.UnmarshalerJSONObject.
-func (v *DeclarationRegistrationOptions) UnmarshalJSONObject(dec *gojay.Decoder, k string) error {
-	switch k {
-	case keyWorkDoneProgress:
-		return dec.Bool(&v.WorkDoneProgress)
-	case keyDocumentSelector:
-		if v.DocumentSelector == nil {
-			v.DocumentSelector = DocumentSelector{}
-		}
-		return dec.Array(&v.DocumentSelector)
-	case keyID:
-		return dec.String(&v.ID)
-	}
-	return nil
-}
-
-// NKeys implements gojay.UnmarshalerJSONObject.
-func (v *DeclarationRegistrationOptions) NKeys() int { return 3 }
-
-// compile time check whether the DeclarationOptions implements a gojay.MarshalerJSONObject and gojay.UnmarshalerJSONObject interfaces.
-var (
-	_ gojay.MarshalerJSONObject   = (*DeclarationOptions)(nil)
-	_ gojay.UnmarshalerJSONObject = (*DeclarationOptions)(nil)
 )
 
 // MarshalJSONObject implements gojay.MarshalerJSONObject.
@@ -851,56 +691,6 @@ var (
 )
 
 // MarshalJSONObject implements gojay.MarshalerJSONObject.
-func (v *DocumentRangeFormattingOptions) MarshalJSONObject(enc *gojay.Encoder) {
-	enc.BoolKeyOmitEmpty(keyWorkDoneProgress, v.WorkDoneProgress)
-}
-
-// IsNil returns wether the structure is nil value or not.
-func (v *DocumentRangeFormattingOptions) IsNil() bool { return v == nil }
-
-// UnmarshalJSONObject implements gojay's UnmarshalerJSONObject.
-func (v *DocumentRangeFormattingOptions) UnmarshalJSONObject(dec *gojay.Decoder, k string) error {
-	if k == keyWorkDoneProgress {
-		return dec.Bool(&v.WorkDoneProgress)
-	}
-	return nil
-}
-
-// NKeys returns the number of keys to unmarshal.
-func (v *DocumentRangeFormattingOptions) NKeys() int { return 1 }
-
-// compile time check whether the DocumentRangeFormattingOptions implements a gojay.MarshalerJSONObject and gojay.UnmarshalerJSONObject interfaces.
-var (
-	_ gojay.MarshalerJSONObject   = (*DocumentRangeFormattingOptions)(nil)
-	_ gojay.UnmarshalerJSONObject = (*DocumentRangeFormattingOptions)(nil)
-)
-
-// MarshalJSONObject implements gojay.MarshalerJSONObject.
-func (v *DefinitionOptions) MarshalJSONObject(enc *gojay.Encoder) {
-	enc.BoolKeyOmitEmpty(keyWorkDoneProgress, v.WorkDoneProgress)
-}
-
-// IsNil implements gojay.MarshalerJSONObject.
-func (v *DefinitionOptions) IsNil() bool { return v == nil }
-
-// UnmarshalJSONObject implements gojay.UnmarshalerJSONObject.
-func (v *DefinitionOptions) UnmarshalJSONObject(dec *gojay.Decoder, k string) error {
-	if k == keyWorkDoneProgress {
-		return dec.Bool(&v.WorkDoneProgress)
-	}
-	return nil
-}
-
-// NKeys implements gojay.UnmarshalerJSONObject.
-func (v *DefinitionOptions) NKeys() int { return 1 }
-
-// compile time check whether the DefinitionOptions implements a gojay.MarshalerJSONObject and gojay.UnmarshalerJSONObject interfaces.
-var (
-	_ gojay.MarshalerJSONObject   = (*DefinitionOptions)(nil)
-	_ gojay.UnmarshalerJSONObject = (*DefinitionOptions)(nil)
-)
-
-// MarshalJSONObject implements gojay.MarshalerJSONObject.
 func (v *DefinitionParams) MarshalJSONObject(enc *gojay.Encoder) {
 	enc.ObjectKeyOmitEmpty(keyTextDocument, &v.TextDocument)
 	enc.ObjectKeyOmitEmpty(keyPosition, &v.Position)
@@ -933,66 +723,6 @@ func (v *DefinitionParams) NKeys() int { return 4 }
 var (
 	_ gojay.MarshalerJSONObject   = (*DefinitionParams)(nil)
 	_ gojay.UnmarshalerJSONObject = (*DefinitionParams)(nil)
-)
-
-// MarshalJSONObject implements gojay.MarshalerJSONObject.
-func (v *TypeDefinitionOptions) MarshalJSONObject(enc *gojay.Encoder) {
-	enc.BoolKeyOmitEmpty(keyWorkDoneProgress, v.WorkDoneProgress)
-}
-
-// IsNil implements gojay.MarshalerJSONObject.
-func (v *TypeDefinitionOptions) IsNil() bool { return v == nil }
-
-// UnmarshalJSONObject implements gojay.UnmarshalerJSONObject.
-func (v *TypeDefinitionOptions) UnmarshalJSONObject(dec *gojay.Decoder, k string) error {
-	if k == keyWorkDoneProgress {
-		return dec.Bool(&v.WorkDoneProgress)
-	}
-	return nil
-}
-
-// NKeys implements gojay.UnmarshalerJSONObject.
-func (v *TypeDefinitionOptions) NKeys() int { return 1 }
-
-// compile time check whether the TypeDefinitionOptions implements a gojay.MarshalerJSONObject and gojay.UnmarshalerJSONObject interfaces.
-var (
-	_ gojay.MarshalerJSONObject   = (*TypeDefinitionOptions)(nil)
-	_ gojay.UnmarshalerJSONObject = (*TypeDefinitionOptions)(nil)
-)
-
-// MarshalJSONObject implements gojay.MarshalerJSONObject.
-func (v *TypeDefinitionRegistrationOptions) MarshalJSONObject(enc *gojay.Encoder) {
-	enc.AddArrayKey(keyDocumentSelector, &v.DocumentSelector)
-	enc.BoolKeyOmitEmpty(keyWorkDoneProgress, v.WorkDoneProgress)
-	enc.StringKeyOmitEmpty(keyID, v.ID)
-}
-
-// IsNil implements gojay.MarshalerJSONObject.
-func (v *TypeDefinitionRegistrationOptions) IsNil() bool { return v == nil }
-
-// UnmarshalJSONObject implements gojay.UnmarshalerJSONObject.
-func (v *TypeDefinitionRegistrationOptions) UnmarshalJSONObject(dec *gojay.Decoder, k string) error {
-	switch k {
-	case keyWorkDoneProgress:
-		return dec.Bool(&v.WorkDoneProgress)
-	case keyDocumentSelector:
-		if v.DocumentSelector == nil {
-			v.DocumentSelector = DocumentSelector{}
-		}
-		return dec.Array(&v.DocumentSelector)
-	case keyID:
-		return dec.String(&v.ID)
-	}
-	return nil
-}
-
-// NKeys implements gojay.UnmarshalerJSONObject.
-func (v *TypeDefinitionRegistrationOptions) NKeys() int { return 3 }
-
-// compile time check whether the TypeDefinitionRegistrationOptions implements a gojay.MarshalerJSONObject and gojay.UnmarshalerJSONObject interfaces.
-var (
-	_ gojay.MarshalerJSONObject   = (*TypeDefinitionRegistrationOptions)(nil)
-	_ gojay.UnmarshalerJSONObject = (*TypeDefinitionRegistrationOptions)(nil)
 )
 
 // MarshalJSONObject implements gojay.MarshalerJSONObject.
@@ -1031,66 +761,6 @@ var (
 )
 
 // MarshalJSONObject implements gojay.MarshalerJSONObject.
-func (v *ImplementationOptions) MarshalJSONObject(enc *gojay.Encoder) {
-	enc.BoolKeyOmitEmpty(keyWorkDoneProgress, v.WorkDoneProgress)
-}
-
-// IsNil implements gojay.MarshalerJSONObject.
-func (v *ImplementationOptions) IsNil() bool { return v == nil }
-
-// UnmarshalJSONObject implements gojay.UnmarshalerJSONObject.
-func (v *ImplementationOptions) UnmarshalJSONObject(dec *gojay.Decoder, k string) error {
-	if k == keyWorkDoneProgress {
-		return dec.Bool(&v.WorkDoneProgress)
-	}
-	return nil
-}
-
-// NKeys implements gojay.UnmarshalerJSONObject.
-func (v *ImplementationOptions) NKeys() int { return 1 }
-
-// compile time check whether the ImplementationOptions implements a gojay.MarshalerJSONObject and gojay.UnmarshalerJSONObject interfaces.
-var (
-	_ gojay.MarshalerJSONObject   = (*ImplementationOptions)(nil)
-	_ gojay.UnmarshalerJSONObject = (*ImplementationOptions)(nil)
-)
-
-// MarshalJSONObject implements gojay.MarshalerJSONObject.
-func (v *ImplementationRegistrationOptions) MarshalJSONObject(enc *gojay.Encoder) {
-	enc.AddArrayKey(keyDocumentSelector, &v.DocumentSelector)
-	enc.BoolKeyOmitEmpty(keyWorkDoneProgress, v.WorkDoneProgress)
-	enc.StringKeyOmitEmpty(keyID, v.ID)
-}
-
-// IsNil implements gojay.MarshalerJSONObject.
-func (v *ImplementationRegistrationOptions) IsNil() bool { return v == nil }
-
-// UnmarshalJSONObject implements gojay.UnmarshalerJSONObject.
-func (v *ImplementationRegistrationOptions) UnmarshalJSONObject(dec *gojay.Decoder, k string) error {
-	switch k {
-	case keyDocumentSelector:
-		if v.DocumentSelector == nil {
-			v.DocumentSelector = DocumentSelector{}
-		}
-		return dec.Array(&v.DocumentSelector)
-	case keyWorkDoneProgress:
-		return dec.Bool(&v.WorkDoneProgress)
-	case keyID:
-		return dec.String(&v.ID)
-	}
-	return nil
-}
-
-// NKeys implements gojay.UnmarshalerJSONObject.
-func (v *ImplementationRegistrationOptions) NKeys() int { return 3 }
-
-// compile time check whether the ImplementationRegistrationOptions implements a gojay.MarshalerJSONObject and gojay.UnmarshalerJSONObject interfaces.
-var (
-	_ gojay.MarshalerJSONObject   = (*ImplementationRegistrationOptions)(nil)
-	_ gojay.UnmarshalerJSONObject = (*ImplementationRegistrationOptions)(nil)
-)
-
-// MarshalJSONObject implements gojay.MarshalerJSONObject.
 func (v *ImplementationParams) MarshalJSONObject(enc *gojay.Encoder) {
 	enc.ObjectKeyOmitEmpty(keyTextDocument, &v.TextDocument)
 	enc.ObjectKeyOmitEmpty(keyPosition, &v.Position)
@@ -1123,95 +793,6 @@ func (v *ImplementationParams) NKeys() int { return 4 }
 var (
 	_ gojay.MarshalerJSONObject   = (*ImplementationParams)(nil)
 	_ gojay.UnmarshalerJSONObject = (*ImplementationParams)(nil)
-)
-
-// CodeActionKinds represents a slice of CodeActionKind.
-type CodeActionKinds []CodeActionKind
-
-// compile time check whether the CodeActionKinds implements a gojay.MarshalerJSONArray and gojay.UnmarshalerJSONArray interfaces.
-var (
-	_ gojay.MarshalerJSONArray   = (*CodeActionKinds)(nil)
-	_ gojay.UnmarshalerJSONArray = (*CodeActionKinds)(nil)
-)
-
-// MarshalJSONArray implements gojay.MarshalerJSONArray.
-func (v CodeActionKinds) MarshalJSONArray(enc *gojay.Encoder) {
-	for i := range v {
-		enc.String(string(v[i]))
-	}
-}
-
-// IsNil implements gojay.MarshalerJSONArray.
-func (v CodeActionKinds) IsNil() bool { return len(v) == 0 }
-
-// UnmarshalJSONArray implements gojay.UnmarshalerJSONArray.
-func (v *CodeActionKinds) UnmarshalJSONArray(dec *gojay.Decoder) error {
-	var value CodeActionKind
-	if err := dec.String((*string)(&value)); err != nil {
-		return err
-	}
-	*v = append(*v, value)
-	return nil
-}
-
-// MarshalJSONObject implements gojay.MarshalerJSONObject.
-func (v *FoldingRangeOptions) MarshalJSONObject(enc *gojay.Encoder) {
-	enc.BoolKeyOmitEmpty(keyWorkDoneProgress, v.WorkDoneProgress)
-}
-
-// IsNil implements gojay.MarshalerJSONObject.
-func (v *FoldingRangeOptions) IsNil() bool { return v == nil }
-
-// UnmarshalJSONObject implements gojay.UnmarshalerJSONObject.
-func (v *FoldingRangeOptions) UnmarshalJSONObject(dec *gojay.Decoder, k string) error {
-	if k == keyWorkDoneProgress {
-		return dec.Bool(&v.WorkDoneProgress)
-	}
-	return nil
-}
-
-// NKeys implements gojay.UnmarshalerJSONObject.
-func (v *FoldingRangeOptions) NKeys() int { return 1 }
-
-// compile time check whether the FoldingRangeOptions implements a gojay.MarshalerJSONObject and gojay.UnmarshalerJSONObject interfaces.
-var (
-	_ gojay.MarshalerJSONObject   = (*FoldingRangeOptions)(nil)
-	_ gojay.UnmarshalerJSONObject = (*FoldingRangeOptions)(nil)
-)
-
-// MarshalJSONObject implements gojay.MarshalerJSONObject.
-func (v *FoldingRangeRegistrationOptions) MarshalJSONObject(enc *gojay.Encoder) {
-	enc.AddArrayKey(keyDocumentSelector, &v.DocumentSelector)
-	enc.BoolKeyOmitEmpty(keyWorkDoneProgress, v.WorkDoneProgress)
-	enc.StringKeyOmitEmpty(keyID, v.ID)
-}
-
-// IsNil implements gojay.MarshalerJSONObject.
-func (v *FoldingRangeRegistrationOptions) IsNil() bool { return v == nil }
-
-// UnmarshalJSONObject implements gojay.UnmarshalerJSONObject.
-func (v *FoldingRangeRegistrationOptions) UnmarshalJSONObject(dec *gojay.Decoder, k string) error {
-	switch k {
-	case keyDocumentSelector:
-		if v.DocumentSelector == nil {
-			v.DocumentSelector = DocumentSelector{}
-		}
-		return dec.Array(&v.DocumentSelector)
-	case keyWorkDoneProgress:
-		return dec.Bool(&v.WorkDoneProgress)
-	case keyID:
-		return dec.String(&v.ID)
-	}
-	return nil
-}
-
-// NKeys implements gojay.UnmarshalerJSONObject.
-func (v *FoldingRangeRegistrationOptions) NKeys() int { return 4 }
-
-// compile time check whether the FoldingRangeRegistrationOptions implements a gojay.MarshalerJSONObject and gojay.UnmarshalerJSONObject interfaces.
-var (
-	_ gojay.MarshalerJSONObject   = (*FoldingRangeRegistrationOptions)(nil)
-	_ gojay.UnmarshalerJSONObject = (*FoldingRangeRegistrationOptions)(nil)
 )
 
 // TokenFormats represents a slice of TokenFormat.
@@ -1393,72 +974,6 @@ var (
 )
 
 // MarshalJSONObject implements gojay.MarshalerJSONObject.
-func (v *CompletionOptions) MarshalJSONObject(enc *gojay.Encoder) {
-	enc.BoolKeyOmitEmpty(keyResolveProvider, v.ResolveProvider)
-	enc.ArrayKeyOmitEmpty(keyTriggerCharacters, (*Strings)(&v.TriggerCharacters))
-}
-
-// IsNil returns wether the structure is nil value or not.
-func (v *CompletionOptions) IsNil() bool { return v == nil }
-
-// UnmarshalJSONObject implements gojay's UnmarshalerJSONObject.
-func (v *CompletionOptions) UnmarshalJSONObject(dec *gojay.Decoder, k string) error {
-	switch k {
-	case keyResolveProvider:
-		return dec.Bool(&v.ResolveProvider)
-	case keyTriggerCharacters:
-		if v.TriggerCharacters == nil {
-			v.TriggerCharacters = Strings{}
-		}
-		return dec.Array((*Strings)(&v.TriggerCharacters))
-	}
-	return nil
-}
-
-// NKeys returns the number of keys to unmarshal.
-func (v *CompletionOptions) NKeys() int { return 2 }
-
-// compile time check whether the CompletionOptions implements a gojay.MarshalerJSONObject and gojay.UnmarshalerJSONObject interfaces.
-var (
-	_ gojay.MarshalerJSONObject   = (*CompletionOptions)(nil)
-	_ gojay.UnmarshalerJSONObject = (*CompletionOptions)(nil)
-)
-
-// MarshalJSONObject implements gojay.MarshalerJSONObject.
-func (v *SignatureHelpOptions) MarshalJSONObject(enc *gojay.Encoder) {
-	enc.ArrayKeyOmitEmpty(keyTriggerCharacters, (*Strings)(&v.TriggerCharacters))
-	enc.ArrayKeyOmitEmpty(keyRetriggerCharacters, (*Strings)(&v.RetriggerCharacters))
-}
-
-// IsNil returns wether the structure is nil value or not.
-func (v *SignatureHelpOptions) IsNil() bool { return v == nil }
-
-// UnmarshalJSONObject implements gojay's UnmarshalerJSONObject.
-func (v *SignatureHelpOptions) UnmarshalJSONObject(dec *gojay.Decoder, k string) error {
-	switch k {
-	case keyTriggerCharacters:
-		var values Strings
-		err := dec.Array(&values)
-		if err == nil && len(values) > 0 {
-			v.TriggerCharacters = []string(values)
-		}
-		return err
-	case keyRetriggerCharacters:
-		return dec.Array((*Strings)(&v.RetriggerCharacters))
-	}
-	return nil
-}
-
-// NKeys returns the number of keys to unmarshal.
-func (v *SignatureHelpOptions) NKeys() int { return 2 }
-
-// compile time check whether the SignatureHelpOptions implements a gojay.MarshalerJSONObject and gojay.UnmarshalerJSONObject interfaces.
-var (
-	_ gojay.MarshalerJSONObject   = (*SignatureHelpOptions)(nil)
-	_ gojay.UnmarshalerJSONObject = (*SignatureHelpOptions)(nil)
-)
-
-// MarshalJSONObject implements gojay.MarshalerJSONObject.
 func (v *ReferencesOptions) MarshalJSONObject(enc *gojay.Encoder) {
 	enc.BoolKeyOmitEmpty(keyWorkDoneProgress, v.WorkDoneProgress)
 }
@@ -1481,380 +996,6 @@ func (v *ReferencesOptions) NKeys() int { return 1 }
 var (
 	_ gojay.MarshalerJSONObject   = (*ReferencesOptions)(nil)
 	_ gojay.UnmarshalerJSONObject = (*ReferencesOptions)(nil)
-)
-
-// MarshalJSONObject implements gojay.MarshalerJSONObject.
-func (v *CodeActionOptions) MarshalJSONObject(enc *gojay.Encoder) {
-	enc.ArrayKeyOmitEmpty(keyCodeActionKinds, (*CodeActionKinds)(&v.CodeActionKinds))
-	enc.BoolKeyOmitEmpty(keyResolveProvider, v.ResolveProvider)
-}
-
-// IsNil returns wether the structure is nil value or not.
-func (v *CodeActionOptions) IsNil() bool { return v == nil }
-
-// UnmarshalJSONObject implements gojay's UnmarshalerJSONObject.
-func (v *CodeActionOptions) UnmarshalJSONObject(dec *gojay.Decoder, k string) error {
-	switch k {
-	case keyCodeActionKinds:
-		return dec.Array((*CodeActionKinds)(&v.CodeActionKinds))
-	case keyResolveProvider:
-		return dec.Bool(&v.ResolveProvider)
-	}
-	return nil
-}
-
-// NKeys returns the number of keys to unmarshal.
-func (v *CodeActionOptions) NKeys() int { return 2 }
-
-// compile time check whether the CodeActionOptions implements a gojay.MarshalerJSONObject and gojay.UnmarshalerJSONObject interfaces.
-var (
-	_ gojay.MarshalerJSONObject   = (*CodeActionOptions)(nil)
-	_ gojay.UnmarshalerJSONObject = (*CodeActionOptions)(nil)
-)
-
-// MarshalJSONObject implements gojay.MarshalerJSONObject.
-func (v *CodeLensOptions) MarshalJSONObject(enc *gojay.Encoder) {
-	enc.BoolKeyOmitEmpty(keyResolveProvider, v.ResolveProvider)
-}
-
-// IsNil returns wether the structure is nil value or not.
-func (v *CodeLensOptions) IsNil() bool { return v == nil }
-
-// UnmarshalJSONObject implements gojay's UnmarshalerJSONObject.
-func (v *CodeLensOptions) UnmarshalJSONObject(dec *gojay.Decoder, k string) error {
-	if k == keyResolveProvider {
-		return dec.Bool(&v.ResolveProvider)
-	}
-	return nil
-}
-
-// NKeys returns the number of keys to unmarshal.
-func (v *CodeLensOptions) NKeys() int { return 1 }
-
-// compile time check whether the CodeLensOptions implements a gojay.MarshalerJSONObject and gojay.UnmarshalerJSONObject interfaces.
-var (
-	_ gojay.MarshalerJSONObject   = (*CodeLensOptions)(nil)
-	_ gojay.UnmarshalerJSONObject = (*CodeLensOptions)(nil)
-)
-
-// MarshalJSONObject implements gojay.MarshalerJSONObject.
-func (v DocumentOnTypeFormattingOptions) MarshalJSONObject(enc *gojay.Encoder) {
-	enc.StringKey(keyFirstTriggerCharacter, v.FirstTriggerCharacter)
-	enc.ArrayKeyOmitEmpty(keyMoreTriggerCharacter, (*Strings)(&v.MoreTriggerCharacter))
-}
-
-// IsNil returns wether the structure is nil value or not.
-func (v *DocumentOnTypeFormattingOptions) IsNil() bool { return v == nil }
-
-// UnmarshalJSONObject implements gojay's UnmarshalerJSONObject.
-func (v *DocumentOnTypeFormattingOptions) UnmarshalJSONObject(dec *gojay.Decoder, k string) error {
-	switch k {
-	case keyFirstTriggerCharacter:
-		return dec.String(&v.FirstTriggerCharacter)
-	case keyMoreTriggerCharacter:
-		var values Strings
-		err := dec.Array(&values)
-		if err == nil && len(values) > 0 {
-			v.MoreTriggerCharacter = []string(values)
-		}
-		return err
-	}
-	return nil
-}
-
-// NKeys returns the number of keys to unmarshal.
-func (v *DocumentOnTypeFormattingOptions) NKeys() int { return 2 }
-
-// compile time check whether the DocumentOnTypeFormattingOptions implements a gojay.MarshalerJSONObject and gojay.UnmarshalerJSONObject interfaces.
-var (
-	_ gojay.MarshalerJSONObject   = (*DocumentOnTypeFormattingOptions)(nil)
-	_ gojay.UnmarshalerJSONObject = (*DocumentOnTypeFormattingOptions)(nil)
-)
-
-// MarshalJSONObject implements gojay.MarshalerJSONObject.
-func (v *RenameOptions) MarshalJSONObject(enc *gojay.Encoder) {
-	enc.BoolKeyOmitEmpty(keyPrepareProvider, v.PrepareProvider)
-}
-
-// IsNil returns wether the structure is nil value or not.
-func (v *RenameOptions) IsNil() bool { return v == nil }
-
-// UnmarshalJSONObject implements gojay's UnmarshalerJSONObject.
-func (v *RenameOptions) UnmarshalJSONObject(dec *gojay.Decoder, k string) error {
-	if k == keyPrepareProvider {
-		return dec.Bool(&v.PrepareProvider)
-	}
-	return nil
-}
-
-// NKeys returns the number of keys to unmarshal.
-func (v *RenameOptions) NKeys() int { return 1 }
-
-// compile time check whether the RenameOptions implements a gojay.MarshalerJSONObject and gojay.UnmarshalerJSONObject interfaces.
-var (
-	_ gojay.MarshalerJSONObject   = (*RenameOptions)(nil)
-	_ gojay.UnmarshalerJSONObject = (*RenameOptions)(nil)
-)
-
-// MarshalJSONObject implements gojay.MarshalerJSONObject.
-func (v *DocumentLinkOptions) MarshalJSONObject(enc *gojay.Encoder) {
-	enc.BoolKeyOmitEmpty(keyResolveProvider, v.ResolveProvider)
-}
-
-// IsNil returns wether the structure is nil value or not.
-func (v *DocumentLinkOptions) IsNil() bool { return v == nil }
-
-// UnmarshalJSONObject implements gojay's UnmarshalerJSONObject.
-func (v *DocumentLinkOptions) UnmarshalJSONObject(dec *gojay.Decoder, k string) error {
-	if k == keyResolveProvider {
-		return dec.Bool(&v.ResolveProvider)
-	}
-	return nil
-}
-
-// NKeys returns the number of keys to unmarshal.
-func (v *DocumentLinkOptions) NKeys() int { return 1 }
-
-// compile time check whether the DocumentLinkOptions implements a gojay.MarshalerJSONObject and gojay.UnmarshalerJSONObject interfaces.
-var (
-	_ gojay.MarshalerJSONObject   = (*DocumentLinkOptions)(nil)
-	_ gojay.UnmarshalerJSONObject = (*DocumentLinkOptions)(nil)
-)
-
-// MarshalJSONObject implements gojay.MarshalerJSONObject.
-func (v *ExecuteCommandOptions) MarshalJSONObject(enc *gojay.Encoder) {
-	enc.ArrayKey(keyCommands, (*Strings)(&v.Commands))
-}
-
-// IsNil returns wether the structure is nil value or not.
-func (v *ExecuteCommandOptions) IsNil() bool { return v == nil }
-
-// UnmarshalJSONObject implements gojay's UnmarshalerJSONObject.
-func (v *ExecuteCommandOptions) UnmarshalJSONObject(dec *gojay.Decoder, k string) error {
-	if k == keyCommands {
-		return dec.Array((*Strings)(&v.Commands))
-	}
-	return nil
-}
-
-// NKeys returns the number of keys to unmarshal.
-func (v *ExecuteCommandOptions) NKeys() int { return 1 }
-
-// compile time check whether the ExecuteCommandOptions implements a gojay.MarshalerJSONObject and gojay.UnmarshalerJSONObject interfaces.
-var (
-	_ gojay.MarshalerJSONObject   = (*ExecuteCommandOptions)(nil)
-	_ gojay.UnmarshalerJSONObject = (*ExecuteCommandOptions)(nil)
-)
-
-// MarshalJSONObject implements gojay.MarshalerJSONObject.
-func (v *SaveOptions) MarshalJSONObject(enc *gojay.Encoder) {
-	enc.BoolKeyOmitEmpty(keyIncludeText, v.IncludeText)
-}
-
-// IsNil returns wether the structure is nil value or not.
-func (v *SaveOptions) IsNil() bool { return v == nil }
-
-// UnmarshalJSONObject implements gojay's UnmarshalerJSONObject.
-func (v *SaveOptions) UnmarshalJSONObject(dec *gojay.Decoder, k string) error {
-	if k == keyIncludeText {
-		return dec.Bool(&v.IncludeText)
-	}
-	return nil
-}
-
-// NKeys returns the number of keys to unmarshal.
-func (v *SaveOptions) NKeys() int { return 1 }
-
-// compile time check whether the SaveOptions implements a gojay.MarshalerJSONObject and gojay.UnmarshalerJSONObject interfaces.
-var (
-	_ gojay.MarshalerJSONObject   = (*SaveOptions)(nil)
-	_ gojay.UnmarshalerJSONObject = (*SaveOptions)(nil)
-)
-
-// MarshalJSONObject implements gojay.MarshalerJSONObject.
-func (v *TextDocumentSyncOptions) MarshalJSONObject(enc *gojay.Encoder) {
-	enc.BoolKeyOmitEmpty(keyOpenClose, v.OpenClose)
-	enc.Float64KeyOmitEmpty(keyChange, float64(v.Change))
-	enc.BoolKeyOmitEmpty(keyWillSave, v.WillSave)
-	enc.BoolKeyOmitEmpty(keyWillSaveWaitUntil, v.WillSaveWaitUntil)
-	enc.ObjectKeyOmitEmpty(keySave, v.Save)
-}
-
-// IsNil returns wether the structure is nil value or not.
-func (v *TextDocumentSyncOptions) IsNil() bool { return v == nil }
-
-// UnmarshalJSONObject implements gojay's UnmarshalerJSONObject.
-func (v *TextDocumentSyncOptions) UnmarshalJSONObject(dec *gojay.Decoder, k string) error {
-	switch k {
-	case keyOpenClose:
-		return dec.Bool(&v.OpenClose)
-	case keyChange:
-		return dec.Float64((*float64)(&v.Change))
-	case keyWillSave:
-		return dec.Bool(&v.WillSave)
-	case keyWillSaveWaitUntil:
-		return dec.Bool(&v.WillSaveWaitUntil)
-	case keySave:
-		if v.Save == nil {
-			v.Save = &SaveOptions{}
-		}
-		return dec.Object(v.Save)
-	}
-	return nil
-}
-
-// NKeys returns the number of keys to unmarshal.
-func (v *TextDocumentSyncOptions) NKeys() int { return 5 }
-
-// compile time check whether the TextDocumentSyncOptions implements a gojay.MarshalerJSONObject and gojay.UnmarshalerJSONObject interfaces.
-var (
-	_ gojay.MarshalerJSONObject   = (*TextDocumentSyncOptions)(nil)
-	_ gojay.UnmarshalerJSONObject = (*TextDocumentSyncOptions)(nil)
-)
-
-// MarshalJSONObject implements gojay.MarshalerJSONObject.
-func (v *HoverOptions) MarshalJSONObject(enc *gojay.Encoder) {
-	enc.BoolKeyOmitEmpty(keyWorkDoneProgress, v.WorkDoneProgress)
-}
-
-// IsNil implements gojay.MarshalerJSONObject.
-func (v *HoverOptions) IsNil() bool { return v == nil }
-
-// UnmarshalJSONObject implements gojay.UnmarshalerJSONObject.
-func (v *HoverOptions) UnmarshalJSONObject(dec *gojay.Decoder, k string) error {
-	if k == keyWorkDoneProgress {
-		return dec.Bool(&v.WorkDoneProgress)
-	}
-	return nil
-}
-
-// NKeys implements gojay.UnmarshalerJSONObject.
-func (v *HoverOptions) NKeys() int { return 1 }
-
-// compile time check whether the HoverOptions implements a gojay.MarshalerJSONObject and gojay.UnmarshalerJSONObject interfaces.
-var (
-	_ gojay.MarshalerJSONObject   = (*HoverOptions)(nil)
-	_ gojay.UnmarshalerJSONObject = (*HoverOptions)(nil)
-)
-
-// MarshalJSONObject implements gojay.MarshalerJSONObject.
-func (v *SemanticTokensOptions) MarshalJSONObject(enc *gojay.Encoder) {
-	enc.BoolKeyOmitEmpty(keyWorkDoneProgress, v.WorkDoneProgress)
-}
-
-// IsNil returns wether the structure is nil value or not.
-func (v *SemanticTokensOptions) IsNil() bool { return v == nil }
-
-// UnmarshalJSONObject implements gojay's UnmarshalerJSONObject.
-func (v *SemanticTokensOptions) UnmarshalJSONObject(dec *gojay.Decoder, k string) error {
-	if k == keyWorkDoneProgress {
-		return dec.Bool(&v.WorkDoneProgress)
-	}
-	return nil
-}
-
-// NKeys returns the number of keys to unmarshal.
-func (v *SemanticTokensOptions) NKeys() int { return 1 }
-
-// compile time check whether the SemanticTokensOptions implements a gojay.MarshalerJSONObject and gojay.UnmarshalerJSONObject interfaces.
-var (
-	_ gojay.MarshalerJSONObject   = (*SemanticTokensOptions)(nil)
-	_ gojay.UnmarshalerJSONObject = (*SemanticTokensOptions)(nil)
-)
-
-// MarshalJSONObject implements gojay.MarshalerJSONObject.
-func (v *SemanticTokensRegistrationOptions) MarshalJSONObject(enc *gojay.Encoder) {
-	enc.AddArrayKey(keyDocumentSelector, &v.DocumentSelector)
-	enc.BoolKeyOmitEmpty(keyWorkDoneProgress, v.WorkDoneProgress)
-	enc.StringKeyOmitEmpty(keyID, v.ID)
-}
-
-// IsNil implements gojay.MarshalerJSONObject.
-func (v *SemanticTokensRegistrationOptions) IsNil() bool { return v == nil }
-
-// UnmarshalJSONObject implements gojay.UnmarshalerJSONObject.
-func (v *SemanticTokensRegistrationOptions) UnmarshalJSONObject(dec *gojay.Decoder, k string) error {
-	switch k {
-	case keyDocumentSelector:
-		if v.DocumentSelector == nil {
-			v.DocumentSelector = DocumentSelector{}
-		}
-		return dec.Array(&v.DocumentSelector)
-	case keyWorkDoneProgress:
-		return dec.Bool(&v.WorkDoneProgress)
-	case keyID:
-		return dec.String(&v.ID)
-	}
-	return nil
-}
-
-// NKeys implements gojay.UnmarshalerJSONObject.
-func (v *SemanticTokensRegistrationOptions) NKeys() int { return 3 }
-
-// compile time check whether the SemanticTokensRegistrationOptions implements a gojay.MarshalerJSONObject and gojay.UnmarshalerJSONObject interfaces.
-var (
-	_ gojay.MarshalerJSONObject   = (*SemanticTokensRegistrationOptions)(nil)
-	_ gojay.UnmarshalerJSONObject = (*SemanticTokensRegistrationOptions)(nil)
-)
-
-// MarshalJSONObject implements gojay.MarshalerJSONObject.
-func (v *LinkedEditingRangeOptions) MarshalJSONObject(enc *gojay.Encoder) {
-	enc.BoolKeyOmitEmpty(keyWorkDoneProgress, v.WorkDoneProgress)
-}
-
-// IsNil returns wether the structure is nil value or not.
-func (v *LinkedEditingRangeOptions) IsNil() bool { return v == nil }
-
-// UnmarshalJSONObject implements gojay's UnmarshalerJSONObject.
-func (v *LinkedEditingRangeOptions) UnmarshalJSONObject(dec *gojay.Decoder, k string) error {
-	if k == keyWorkDoneProgress {
-		return dec.Bool(&v.WorkDoneProgress)
-	}
-	return nil
-}
-
-// NKeys returns the number of keys to unmarshal.
-func (v *LinkedEditingRangeOptions) NKeys() int { return 1 }
-
-// compile time check whether the LinkedEditingRangeOptions implements a gojay.MarshalerJSONObject and gojay.UnmarshalerJSONObject interfaces.
-var (
-	_ gojay.MarshalerJSONObject   = (*LinkedEditingRangeOptions)(nil)
-	_ gojay.UnmarshalerJSONObject = (*LinkedEditingRangeOptions)(nil)
-)
-
-// MarshalJSONObject implements gojay.MarshalerJSONObject.
-func (v *LinkedEditingRangeRegistrationOptions) MarshalJSONObject(enc *gojay.Encoder) {
-	enc.AddArrayKey(keyDocumentSelector, &v.DocumentSelector)
-	enc.BoolKeyOmitEmpty(keyWorkDoneProgress, v.WorkDoneProgress)
-	enc.StringKeyOmitEmpty(keyID, v.ID)
-}
-
-// IsNil implements gojay.MarshalerJSONObject.
-func (v *LinkedEditingRangeRegistrationOptions) IsNil() bool { return v == nil }
-
-// UnmarshalJSONObject implements gojay.UnmarshalerJSONObject.
-func (v *LinkedEditingRangeRegistrationOptions) UnmarshalJSONObject(dec *gojay.Decoder, k string) error {
-	switch k {
-	case keyDocumentSelector:
-		if v.DocumentSelector == nil {
-			v.DocumentSelector = DocumentSelector{}
-		}
-		return dec.Array(&v.DocumentSelector)
-	case keyWorkDoneProgress:
-		return dec.Bool(&v.WorkDoneProgress)
-	case keyID:
-		return dec.String(&v.ID)
-	}
-	return nil
-}
-
-// NKeys implements gojay.UnmarshalerJSONObject.
-func (v *LinkedEditingRangeRegistrationOptions) NKeys() int { return 3 }
-
-// compile time check whether the LinkedEditingRangeRegistrationOptions implements a gojay.MarshalerJSONObject and gojay.UnmarshalerJSONObject interfaces.
-var (
-	_ gojay.MarshalerJSONObject   = (*LinkedEditingRangeRegistrationOptions)(nil)
-	_ gojay.UnmarshalerJSONObject = (*LinkedEditingRangeRegistrationOptions)(nil)
 )
 
 // MarshalJSONObject implements gojay.MarshalerJSONObject.
@@ -1916,63 +1057,6 @@ func (v *LinkedEditingRanges) NKeys() int { return 2 }
 var (
 	_ gojay.MarshalerJSONObject   = (*LinkedEditingRanges)(nil)
 	_ gojay.UnmarshalerJSONObject = (*LinkedEditingRanges)(nil)
-)
-
-// MarshalJSONObject implements gojay.MarshalerJSONObject.
-func (v *MonikerOptions) MarshalJSONObject(enc *gojay.Encoder) {
-	enc.BoolKeyOmitEmpty(keyWorkDoneProgress, v.WorkDoneProgress)
-}
-
-// IsNil returns wether the structure is nil value or not.
-func (v *MonikerOptions) IsNil() bool { return v == nil }
-
-// UnmarshalJSONObject implements gojay's UnmarshalerJSONObject.
-func (v *MonikerOptions) UnmarshalJSONObject(dec *gojay.Decoder, k string) error {
-	if k == keyWorkDoneProgress {
-		return dec.Bool(&v.WorkDoneProgress)
-	}
-	return nil
-}
-
-// NKeys returns the number of keys to unmarshal.
-func (v *MonikerOptions) NKeys() int { return 1 }
-
-// compile time check whether the MonikerOptions implements a gojay.MarshalerJSONObject and gojay.UnmarshalerJSONObject interfaces.
-var (
-	_ gojay.MarshalerJSONObject   = (*MonikerOptions)(nil)
-	_ gojay.UnmarshalerJSONObject = (*MonikerOptions)(nil)
-)
-
-// MarshalJSONObject implements gojay.MarshalerJSONObject.
-func (v *MonikerRegistrationOptions) MarshalJSONObject(enc *gojay.Encoder) {
-	enc.AddArrayKey(keyDocumentSelector, &v.DocumentSelector)
-	enc.BoolKeyOmitEmpty(keyWorkDoneProgress, v.WorkDoneProgress)
-}
-
-// IsNil implements gojay.MarshalerJSONObject.
-func (v *MonikerRegistrationOptions) IsNil() bool { return v == nil }
-
-// UnmarshalJSONObject implements gojay.UnmarshalerJSONObject.
-func (v *MonikerRegistrationOptions) UnmarshalJSONObject(dec *gojay.Decoder, k string) error {
-	switch k {
-	case keyDocumentSelector:
-		if v.DocumentSelector == nil {
-			v.DocumentSelector = DocumentSelector{}
-		}
-		return dec.Array(&v.DocumentSelector)
-	case keyWorkDoneProgress:
-		return dec.Bool(&v.WorkDoneProgress)
-	}
-	return nil
-}
-
-// NKeys implements gojay.UnmarshalerJSONObject.
-func (v *MonikerRegistrationOptions) NKeys() int { return 2 }
-
-// compile time check whether the MonikerRegistrationOptions implements a gojay.MarshalerJSONObject and gojay.UnmarshalerJSONObject interfaces.
-var (
-	_ gojay.MarshalerJSONObject   = (*MonikerRegistrationOptions)(nil)
-	_ gojay.UnmarshalerJSONObject = (*MonikerRegistrationOptions)(nil)
 )
 
 // MarshalJSONObject implements gojay.MarshalerJSONObject.
